+++ Tests/run-unit.php
@@ -97,6 +97,8 @@ namespace {
         \FacturaScripts\Test\Plugins\BeplyPDFStudio\BeplyPdfConfigValidatorTest::class,
         \FacturaScripts\Test\Plugins\BeplyPDFStudio\BeplyPdfCorporateHeaderGeometryTest::class,
         \FacturaScripts\Test\Plugins\BeplyPDFStudio\BeplyPdfDocumentExtensionRegistryTest::class,
+        \FacturaScripts\Test\Plugins\BeplyPDFStudio\BeplyPdfDocumentTotalsConsistencyTest::class,
+        \FacturaScripts\Test\Plugins\BeplyPDFStudio\BeplyPdfInconsistentDocumentGuardTest::class,
         \FacturaScripts\Test\Plugins\BeplyPDFStudio\BeplyPdfDocumentCacheServiceTest::class,
         \FacturaScripts\Test\Plugins\BeplyPDFStudio\BeplyPdfGenericReportBufferTest::class,
         \FacturaScripts\Test\Plugins\BeplyPDFStudio\BeplyPdfInternalFormatPolicyTest::class,